[TWINFACES-547] refactor: rename twin-classes folder
